agency-crm/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
flask-migrate==4.0.5
gunicorn==21.2.0
python-dateutil==2.8.2
openpyxl==3.1.2
requests==2.32.5